third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_character_09.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'character'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} character`
  );
}

function makeSample(offset){
  const kSample = 'Lorem <div  dir="rtl">ipsum dolor sit</div> amett';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right character
testIt('ltr', 0, 'right', '^L|orem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 1, 'right', 'L^o|rem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 2, 'right', 'Lo^r|em <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 3, 'right', 'Lor^e|m <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 4, 'right', 'Lore^m| <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 5, 'right', 'Lorem ^<div dir="rtl">|ipsum dolor sit</div> amett');
testIt('ltr', 6, 'right', 'Lorem ^<div dir="rtl">|ipsum dolor sit</div> amett');
testIt('ltr', 7, 'right', 'Lorem <^d|iv  dir="rtl">ipsum dolor sit amett');
testIt('ltr', 8, 'right', 'Lorem <d|iv dir="rtl">ipsum dolor sit amett</d|iv>');
testIt('ltr', 9, 'right', 'Lorem <di|v dir="rtl">ipsum dolor sit amett</di|v>');
testIt('ltr', 10, 'right', 'Lorem <div| dir="rtl">ipsum dolor sit amett</div|>');
testIt('ltr', 11, 'right', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 12, 'right', 'Lorem <div |dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 13, 'right', 'Lorem <div d|ir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 14, 'right', 'Lorem <div di|r="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 15, 'right', 'Lorem <div dir|="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 16, 'right', 'Lorem <div dir="|&quot;rtl&quot;">ipsum dolor sit</div> amett');
testIt('ltr', 17, 'right', 'Lorem <div dir="|rtl">ipsum dolor sit</div> amett');
testIt('ltr', 18, 'right', 'Lorem <div dir="r|tl">ipsum dolor sit</div> amett');
testIt('ltr', 19, 'right', 'Lorem <div dir="rt|l">ipsum dolor sit</div> amett');
testIt('ltr', 20, 'right', 'Lorem <div dir="rtl|">ipsum dolor sit</div> amett');
testIt('ltr', 21, 'right', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 22, 'right', 'Lorem |<div dir="rtl">^ipsum dolor sit</div> amett');
testIt('ltr', 23, 'right', 'Lorem <div dir="rtl">|i^psum dolor sit</div> amett');
testIt('ltr', 24, 'right', 'Lorem <div dir="rtl">i|p^sum dolor sit</div> amett');
testIt('ltr', 25, 'right', 'Lorem <div dir="rtl">ip|s^um dolor sit</div> amett');
testIt('ltr', 26, 'right', 'Lorem <div dir="rtl">ips|u^m dolor sit</div> amett');
testIt('ltr', 27, 'right', 'Lorem <div dir="rtl">ipsu|m^ dolor sit</div> amett');
testIt('ltr', 28, 'right', 'Lorem <div dir="rtl">ipsum| ^dolor sit</div> amett');
testIt('ltr', 29, 'right', 'Lorem <div dir="rtl">ipsum |d^olor sit</div> amett');
testIt('ltr', 30, 'right', 'Lorem <div dir="rtl">ipsum d|o^lor sit</div> amett');
testIt('ltr', 31, 'right', 'Lorem <div dir="rtl">ipsum do|l^or sit</div> amett');
testIt('ltr', 32, 'right', 'Lorem <div dir="rtl">ipsum dol|o^r sit</div> amett');
testIt('ltr', 33, 'right', 'Lorem <div dir="rtl">ipsum dolo|r^ sit</div> amett');
testIt('ltr', 34, 'right', 'Lorem <div dir="rtl">ipsum dolor| ^sit</div> amett');
testIt('ltr', 35, 'right', 'Lorem <div dir="rtl">ipsum dolor |s^it</div> amett');
testIt('ltr', 36, 'right', 'Lorem <div dir="rtl">ipsum dolor s|i^t</div> amett');
testIt('ltr', 37, 'right', 'Lorem <div dir="rtl">ipsum dolor si|t^</div> amett');
testIt('ltr', 38, 'right', 'Lorem <div dir="rtl">ipsum dolor sit|<^/div> amett</div>');
testIt('ltr', 39, 'right', 'Lorem <div dir="rtl">ipsum dolor si|t^div amett</div>');
testIt('ltr', 40, 'right', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('ltr', 41, 'right', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('ltr', 42, 'right', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('ltr', 43, 'right', 'Lorem <div dir="rtl">ipsum dolor sit</div> ^a|mett');
testIt('ltr', 44, 'right', 'Lorem <div dir="rtl">ipsum dolor sit</div> ^a|mett');
testIt('ltr', 45, 'right', 'Lorem <div dir="rtl">ipsum dolor sit</div> a^m|ett');
testIt('ltr', 46, 'right', 'Lorem <div dir="rtl">ipsum dolor sit</div> am^e|tt');
testIt('ltr', 47, 'right', 'Lorem <div dir="rtl">ipsum dolor sit</div> ame^t|t');
testIt('ltr', 48, 'right', 'Lorem <div dir="rtl">ipsum dolor sit</div> amet^t|');

// ltr left character
testIt('ltr', 0, 'left', '|Lorem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 1, 'left', '|L^orem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 2, 'left', 'L|o^rem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 3, 'left', 'Lo|r^em <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 4, 'left', 'Lor|e^m <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 5, 'left', 'Lore|m^ <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 6, 'left', 'Lore|m^ <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 7, 'left', 'Lorem |<^div  dir="rtl">ipsum dolor sit amett');
testIt('ltr', 8, 'left', 'Lorem <d|iv dir="rtl">ipsum dolor sit amett</d|iv>');
testIt('ltr', 9, 'left', 'Lorem <di|v dir="rtl">ipsum dolor sit amett</di|v>');
testIt('ltr', 10, 'left', 'Lorem <div| dir="rtl">ipsum dolor sit amett</div|>');
testIt('ltr', 11, 'left', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 12, 'left', 'Lorem <div |dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 13, 'left', 'Lorem <div d|ir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 14, 'left', 'Lorem <div di|r="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 15, 'left', 'Lorem <div dir|="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 16, 'left', 'Lorem <div dir="|&quot;rtl&quot;">ipsum dolor sit</div> amett');
testIt('ltr', 17, 'left', 'Lorem <div dir="|rtl">ipsum dolor sit</div> amett');
testIt('ltr', 18, 'left', 'Lorem <div dir="r|tl">ipsum dolor sit</div> amett');
testIt('ltr', 19, 'left', 'Lorem <div dir="rt|l">ipsum dolor sit</div> amett');
testIt('ltr', 20, 'left', 'Lorem <div dir="rtl|">ipsum dolor sit</div> amett');
testIt('ltr', 21, 'left', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 22, 'left', 'Lorem <div dir="rtl">^i|psum dolor sit</div> amett');
testIt('ltr', 23, 'left', 'Lorem <div dir="rtl">i^p|sum dolor sit</div> amett');
testIt('ltr', 24, 'left', 'Lorem <div dir="rtl">ip^s|um dolor sit</div> amett');
testIt('ltr', 25, 'left', 'Lorem <div dir="rtl">ips^u|m dolor sit</div> amett');
testIt('ltr', 26, 'left', 'Lorem <div dir="rtl">ipsu^m| dolor sit</div> amett');
testIt('ltr', 27, 'left', 'Lorem <div dir="rtl">ipsum^ |dolor sit</div> amett');
testIt('ltr', 28, 'left', 'Lorem <div dir="rtl">ipsum ^d|olor sit</div> amett');
testIt('ltr', 29, 'left', 'Lorem <div dir="rtl">ipsum d^o|lor sit</div> amett');
testIt('ltr', 30, 'left', 'Lorem <div dir="rtl">ipsum do^l|or sit</div> amett');
testIt('ltr', 31, 'left', 'Lorem <div dir="rtl">ipsum dol^o|r sit</div> amett');
testIt('ltr', 32, 'left', 'Lorem <div dir="rtl">ipsum dolo^r| sit</div> amett');
testIt('ltr', 33, 'left', 'Lorem <div dir="rtl">ipsum dolor^ |sit</div> amett');
testIt('ltr', 34, 'left', 'Lorem <div dir="rtl">ipsum dolor ^s|it</div> amett');
testIt('ltr', 35, 'left', 'Lorem <div dir="rtl">ipsum dolor s^i|t</div> amett');
testIt('ltr', 36, 'left', 'Lorem <div dir="rtl">ipsum dolor si^t|</div> amett');
testIt('ltr', 37, 'left', 'Lorem <div dir="rtl">ipsum dolor sit^</div>| amett');
testIt('ltr', 38, 'left', 'Lorem <div dir="rtl">ipsum dolor sit<^/|div> amett</div>');
testIt('ltr', 39, 'left', 'Lorem <div dir="rtl">ipsum dolor sitdiv^ |amett</div>');
testIt('ltr', 40, 'left', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('ltr', 41, 'left', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('ltr', 42, 'left', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('ltr', 43, 'left', 'Lorem <div dir="rtl">ipsum dolor sit|</div>^ amett');
testIt('ltr', 44, 'left', 'Lorem <div dir="rtl">ipsum dolor sit|</div>^ amett');
testIt('ltr', 45, 'left', 'Lorem <div dir="rtl">ipsum dolor sit</div> |a^mett');
testIt('ltr', 46, 'left', 'Lorem <div dir="rtl">ipsum dolor sit</div> a|m^ett');
testIt('ltr', 47, 'left', 'Lorem <div dir="rtl">ipsum dolor sit</div> am|e^tt');
testIt('ltr', 48, 'left', 'Lorem <div dir="rtl">ipsum dolor sit</div> ame|t^t');

// ltr forward character
testIt('ltr', 0, 'forward', '^L|orem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 1, 'forward', 'L^o|rem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 2, 'forward', 'Lo^r|em <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 3, 'forward', 'Lor^e|m <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 4, 'forward', 'Lore^m| <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 5, 'forward', 'Lorem ^<div dir="rtl">|ipsum dolor sit</div> amett');
testIt('ltr', 6, 'forward', 'Lorem ^<div dir="rtl">|ipsum dolor sit</div> amett');
testIt('ltr', 7, 'forward', 'Lorem <^d|iv  dir="rtl">ipsum dolor sit amett');
testIt('ltr', 8, 'forward', 'Lorem <d|iv dir="rtl">ipsum dolor sit amett</d|iv>');
testIt('ltr', 9, 'forward', 'Lorem <di|v dir="rtl">ipsum dolor sit amett</di|v>');
testIt('ltr', 10, 'forward', 'Lorem <div| dir="rtl">ipsum dolor sit amett</div|>');
testIt('ltr', 11, 'forward', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 12, 'forward', 'Lorem <div |dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 13, 'forward', 'Lorem <div d|ir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 14, 'forward', 'Lorem <div di|r="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 15, 'forward', 'Lorem <div dir|="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 16, 'forward', 'Lorem <div dir="|&quot;rtl&quot;">ipsum dolor sit</div> amett');
testIt('ltr', 17, 'forward', 'Lorem <div dir="|rtl">ipsum dolor sit</div> amett');
testIt('ltr', 18, 'forward', 'Lorem <div dir="r|tl">ipsum dolor sit</div> amett');
testIt('ltr', 19, 'forward', 'Lorem <div dir="rt|l">ipsum dolor sit</div> amett');
testIt('ltr', 20, 'forward', 'Lorem <div dir="rtl|">ipsum dolor sit</div> amett');
testIt('ltr', 21, 'forward', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 22, 'forward', 'Lorem <div dir="rtl">^i|psum dolor sit</div> amett');
testIt('ltr', 23, 'forward', 'Lorem <div dir="rtl">i^p|sum dolor sit</div> amett');
testIt('ltr', 24, 'forward', 'Lorem <div dir="rtl">ip^s|um dolor sit</div> amett');
testIt('ltr', 25, 'forward', 'Lorem <div dir="rtl">ips^u|m dolor sit</div> amett');
testIt('ltr', 26, 'forward', 'Lorem <div dir="rtl">ipsu^m| dolor sit</div> amett');
testIt('ltr', 27, 'forward', 'Lorem <div dir="rtl">ipsum^ |dolor sit</div> amett');
testIt('ltr', 28, 'forward', 'Lorem <div dir="rtl">ipsum ^d|olor sit</div> amett');
testIt('ltr', 29, 'forward', 'Lorem <div dir="rtl">ipsum d^o|lor sit</div> amett');
testIt('ltr', 30, 'forward', 'Lorem <div dir="rtl">ipsum do^l|or sit</div> amett');
testIt('ltr', 31, 'forward', 'Lorem <div dir="rtl">ipsum dol^o|r sit</div> amett');
testIt('ltr', 32, 'forward', 'Lorem <div dir="rtl">ipsum dolo^r| sit</div> amett');
testIt('ltr', 33, 'forward', 'Lorem <div dir="rtl">ipsum dolor^ |sit</div> amett');
testIt('ltr', 34, 'forward', 'Lorem <div dir="rtl">ipsum dolor ^s|it</div> amett');
testIt('ltr', 35, 'forward', 'Lorem <div dir="rtl">ipsum dolor s^i|t</div> amett');
testIt('ltr', 36, 'forward', 'Lorem <div dir="rtl">ipsum dolor si^t|</div> amett');
testIt('ltr', 37, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit^</div>| amett');
testIt('ltr', 38, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit<^/|div> amett</div>');
testIt('ltr', 39, 'forward', 'Lorem <div dir="rtl">ipsum dolor sitdiv^ |amett</div>');
testIt('ltr', 40, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('ltr', 41, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('ltr', 42, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('ltr', 43, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit</div> ^a|mett');
testIt('ltr', 44, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit</div> ^a|mett');
testIt('ltr', 45, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit</div> a^m|ett');
testIt('ltr', 46, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit</div> am^e|tt');
testIt('ltr', 47, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit</div> ame^t|t');
testIt('ltr', 48, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit</div> amet^t|');

// ltr backward character
testIt('ltr', 0, 'backward', '|Lorem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 1, 'backward', '|L^orem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 2, 'backward', 'L|o^rem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 3, 'backward', 'Lo|r^em <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 4, 'backward', 'Lor|e^m <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 5, 'backward', 'Lore|m^ <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 6, 'backward', 'Lore|m^ <div dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 7, 'backward', 'Lorem |<^div  dir="rtl">ipsum dolor sit amett');
testIt('ltr', 8, 'backward', 'Lorem <d|iv dir="rtl">ipsum dolor sit amett</d|iv>');
testIt('ltr', 9, 'backward', 'Lorem <di|v dir="rtl">ipsum dolor sit amett</di|v>');
testIt('ltr', 10, 'backward', 'Lorem <div| dir="rtl">ipsum dolor sit amett</div|>');
testIt('ltr', 11, 'backward', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 12, 'backward', 'Lorem <div |dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 13, 'backward', 'Lorem <div d|ir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 14, 'backward', 'Lorem <div di|r="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 15, 'backward', 'Lorem <div dir|="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 16, 'backward', 'Lorem <div dir="|&quot;rtl&quot;">ipsum dolor sit</div> amett');
testIt('ltr', 17, 'backward', 'Lorem <div dir="|rtl">ipsum dolor sit</div> amett');
testIt('ltr', 18, 'backward', 'Lorem <div dir="r|tl">ipsum dolor sit</div> amett');
testIt('ltr', 19, 'backward', 'Lorem <div dir="rt|l">ipsum dolor sit</div> amett');
testIt('ltr', 20, 'backward', 'Lorem <div dir="rtl|">ipsum dolor sit</div> amett');
testIt('ltr', 21, 'backward', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('ltr', 22, 'backward', 'Lorem |<div dir="rtl">^ipsum dolor sit</div> amett');
testIt('ltr', 23, 'backward', 'Lorem <div dir="rtl">|i^psum dolor sit</div> amett');
testIt('ltr', 24, 'backward', 'Lorem <div dir="rtl">i|p^sum dolor sit</div> amett');
testIt('ltr', 25, 'backward', 'Lorem <div dir="rtl">ip|s^um dolor sit</div> amett');
testIt('ltr', 26, 'backward', 'Lorem <div dir="rtl">ips|u^m dolor sit</div> amett');
testIt('ltr', 27, 'backward', 'Lorem <div dir="rtl">ipsu|m^ dolor sit</div> amett');
testIt('ltr', 28, 'backward', 'Lorem <div dir="rtl">ipsum| ^dolor sit</div> amett');
testIt('ltr', 29, 'backward', 'Lorem <div dir="rtl">ipsum |d^olor sit</div> amett');
testIt('ltr', 30, 'backward', 'Lorem <div dir="rtl">ipsum d|o^lor sit</div> amett');
testIt('ltr', 31, 'backward', 'Lorem <div dir="rtl">ipsum do|l^or sit</div> amett');
testIt('ltr', 32, 'backward', 'Lorem <div dir="rtl">ipsum dol|o^r sit</div> amett');
testIt('ltr', 33, 'backward', 'Lorem <div dir="rtl">ipsum dolo|r^ sit</div> amett');
testIt('ltr', 34, 'backward', 'Lorem <div dir="rtl">ipsum dolor| ^sit</div> amett');
testIt('ltr', 35, 'backward', 'Lorem <div dir="rtl">ipsum dolor |s^it</div> amett');
testIt('ltr', 36, 'backward', 'Lorem <div dir="rtl">ipsum dolor s|i^t</div> amett');
testIt('ltr', 37, 'backward', 'Lorem <div dir="rtl">ipsum dolor si|t^</div> amett');
testIt('ltr', 38, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit|<^/div> amett</div>');
testIt('ltr', 39, 'backward', 'Lorem <div dir="rtl">ipsum dolor si|t^div amett</div>');
testIt('ltr', 40, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('ltr', 41, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('ltr', 42, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('ltr', 43, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit|</div>^ amett');
testIt('ltr', 44, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit|</div>^ amett');
testIt('ltr', 45, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit</div> |a^mett');
testIt('ltr', 46, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit</div> a|m^ett');
testIt('ltr', 47, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit</div> am|e^tt');
testIt('ltr', 48, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit</div> ame|t^t');

// rtl right character
testIt('rtl', 0, 'right', '^L|orem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 1, 'right', 'L^o|rem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 2, 'right', 'Lo^r|em <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 3, 'right', 'Lor^e|m <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 4, 'right', 'Lore^m| <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 5, 'right', 'Lorem ^<div dir="rtl">|ipsum dolor sit</div> amett');
testIt('rtl', 6, 'right', 'Lorem ^<div dir="rtl">|ipsum dolor sit</div> amett');
testIt('rtl', 7, 'right', 'Lorem <^d|iv  dir="rtl">ipsum dolor sit amett');
testIt('rtl', 8, 'right', 'Lorem <d|iv dir="rtl">ipsum dolor sit amett</d|iv>');
testIt('rtl', 9, 'right', 'Lorem <di|v dir="rtl">ipsum dolor sit amett</di|v>');
testIt('rtl', 10, 'right', 'Lorem <div| dir="rtl">ipsum dolor sit amett</div|>');
testIt('rtl', 11, 'right', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 12, 'right', 'Lorem <div |dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 13, 'right', 'Lorem <div d|ir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 14, 'right', 'Lorem <div di|r="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 15, 'right', 'Lorem <div dir|="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 16, 'right', 'Lorem <div dir="|&quot;rtl&quot;">ipsum dolor sit</div> amett');
testIt('rtl', 17, 'right', 'Lorem <div dir="|rtl">ipsum dolor sit</div> amett');
testIt('rtl', 18, 'right', 'Lorem <div dir="r|tl">ipsum dolor sit</div> amett');
testIt('rtl', 19, 'right', 'Lorem <div dir="rt|l">ipsum dolor sit</div> amett');
testIt('rtl', 20, 'right', 'Lorem <div dir="rtl|">ipsum dolor sit</div> amett');
testIt('rtl', 21, 'right', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 22, 'right', 'Lorem |<div dir="rtl">^ipsum dolor sit</div> amett');
testIt('rtl', 23, 'right', 'Lorem <div dir="rtl">|i^psum dolor sit</div> amett');
testIt('rtl', 24, 'right', 'Lorem <div dir="rtl">i|p^sum dolor sit</div> amett');
testIt('rtl', 25, 'right', 'Lorem <div dir="rtl">ip|s^um dolor sit</div> amett');
testIt('rtl', 26, 'right', 'Lorem <div dir="rtl">ips|u^m dolor sit</div> amett');
testIt('rtl', 27, 'right', 'Lorem <div dir="rtl">ipsu|m^ dolor sit</div> amett');
testIt('rtl', 28, 'right', 'Lorem <div dir="rtl">ipsum| ^dolor sit</div> amett');
testIt('rtl', 29, 'right', 'Lorem <div dir="rtl">ipsum |d^olor sit</div> amett');
testIt('rtl', 30, 'right', 'Lorem <div dir="rtl">ipsum d|o^lor sit</div> amett');
testIt('rtl', 31, 'right', 'Lorem <div dir="rtl">ipsum do|l^or sit</div> amett');
testIt('rtl', 32, 'right', 'Lorem <div dir="rtl">ipsum dol|o^r sit</div> amett');
testIt('rtl', 33, 'right', 'Lorem <div dir="rtl">ipsum dolo|r^ sit</div> amett');
testIt('rtl', 34, 'right', 'Lorem <div dir="rtl">ipsum dolor| ^sit</div> amett');
testIt('rtl', 35, 'right', 'Lorem <div dir="rtl">ipsum dolor |s^it</div> amett');
testIt('rtl', 36, 'right', 'Lorem <div dir="rtl">ipsum dolor s|i^t</div> amett');
testIt('rtl', 37, 'right', 'Lorem <div dir="rtl">ipsum dolor si|t^</div> amett');
testIt('rtl', 38, 'right', 'Lorem <div dir="rtl">ipsum dolor sit|<^/div> amett</div>');
testIt('rtl', 39, 'right', 'Lorem <div dir="rtl">ipsum dolor si|t^div amett</div>');
testIt('rtl', 40, 'right', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('rtl', 41, 'right', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('rtl', 42, 'right', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('rtl', 43, 'right', 'Lorem <div dir="rtl">ipsum dolor sit</div> ^a|mett');
testIt('rtl', 44, 'right', 'Lorem <div dir="rtl">ipsum dolor sit</div> ^a|mett');
testIt('rtl', 45, 'right', 'Lorem <div dir="rtl">ipsum dolor sit</div> a^m|ett');
testIt('rtl', 46, 'right', 'Lorem <div dir="rtl">ipsum dolor sit</div> am^e|tt');
testIt('rtl', 47, 'right', 'Lorem <div dir="rtl">ipsum dolor sit</div> ame^t|t');
testIt('rtl', 48, 'right', 'Lorem <div dir="rtl">ipsum dolor sit</div> amet^t|');

// rtl left character
testIt('rtl', 0, 'left', '|Lorem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 1, 'left', '|L^orem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 2, 'left', 'L|o^rem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 3, 'left', 'Lo|r^em <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 4, 'left', 'Lor|e^m <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 5, 'left', 'Lore|m^ <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 6, 'left', 'Lore|m^ <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 7, 'left', 'Lorem |<^div  dir="rtl">ipsum dolor sit amett');
testIt('rtl', 8, 'left', 'Lorem <d|iv dir="rtl">ipsum dolor sit amett</d|iv>');
testIt('rtl', 9, 'left', 'Lorem <di|v dir="rtl">ipsum dolor sit amett</di|v>');
testIt('rtl', 10, 'left', 'Lorem <div| dir="rtl">ipsum dolor sit amett</div|>');
testIt('rtl', 11, 'left', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 12, 'left', 'Lorem <div |dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 13, 'left', 'Lorem <div d|ir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 14, 'left', 'Lorem <div di|r="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 15, 'left', 'Lorem <div dir|="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 16, 'left', 'Lorem <div dir="|&quot;rtl&quot;">ipsum dolor sit</div> amett');
testIt('rtl', 17, 'left', 'Lorem <div dir="|rtl">ipsum dolor sit</div> amett');
testIt('rtl', 18, 'left', 'Lorem <div dir="r|tl">ipsum dolor sit</div> amett');
testIt('rtl', 19, 'left', 'Lorem <div dir="rt|l">ipsum dolor sit</div> amett');
testIt('rtl', 20, 'left', 'Lorem <div dir="rtl|">ipsum dolor sit</div> amett');
testIt('rtl', 21, 'left', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 22, 'left', 'Lorem <div dir="rtl">^i|psum dolor sit</div> amett');
testIt('rtl', 23, 'left', 'Lorem <div dir="rtl">i^p|sum dolor sit</div> amett');
testIt('rtl', 24, 'left', 'Lorem <div dir="rtl">ip^s|um dolor sit</div> amett');
testIt('rtl', 25, 'left', 'Lorem <div dir="rtl">ips^u|m dolor sit</div> amett');
testIt('rtl', 26, 'left', 'Lorem <div dir="rtl">ipsu^m| dolor sit</div> amett');
testIt('rtl', 27, 'left', 'Lorem <div dir="rtl">ipsum^ |dolor sit</div> amett');
testIt('rtl', 28, 'left', 'Lorem <div dir="rtl">ipsum ^d|olor sit</div> amett');
testIt('rtl', 29, 'left', 'Lorem <div dir="rtl">ipsum d^o|lor sit</div> amett');
testIt('rtl', 30, 'left', 'Lorem <div dir="rtl">ipsum do^l|or sit</div> amett');
testIt('rtl', 31, 'left', 'Lorem <div dir="rtl">ipsum dol^o|r sit</div> amett');
testIt('rtl', 32, 'left', 'Lorem <div dir="rtl">ipsum dolo^r| sit</div> amett');
testIt('rtl', 33, 'left', 'Lorem <div dir="rtl">ipsum dolor^ |sit</div> amett');
testIt('rtl', 34, 'left', 'Lorem <div dir="rtl">ipsum dolor ^s|it</div> amett');
testIt('rtl', 35, 'left', 'Lorem <div dir="rtl">ipsum dolor s^i|t</div> amett');
testIt('rtl', 36, 'left', 'Lorem <div dir="rtl">ipsum dolor si^t|</div> amett');
testIt('rtl', 37, 'left', 'Lorem <div dir="rtl">ipsum dolor sit^</div>| amett');
testIt('rtl', 38, 'left', 'Lorem <div dir="rtl">ipsum dolor sit<^/|div> amett</div>');
testIt('rtl', 39, 'left', 'Lorem <div dir="rtl">ipsum dolor sitdiv^ |amett</div>');
testIt('rtl', 40, 'left', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('rtl', 41, 'left', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('rtl', 42, 'left', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('rtl', 43, 'left', 'Lorem <div dir="rtl">ipsum dolor sit|</div>^ amett');
testIt('rtl', 44, 'left', 'Lorem <div dir="rtl">ipsum dolor sit|</div>^ amett');
testIt('rtl', 45, 'left', 'Lorem <div dir="rtl">ipsum dolor sit</div> |a^mett');
testIt('rtl', 46, 'left', 'Lorem <div dir="rtl">ipsum dolor sit</div> a|m^ett');
testIt('rtl', 47, 'left', 'Lorem <div dir="rtl">ipsum dolor sit</div> am|e^tt');
testIt('rtl', 48, 'left', 'Lorem <div dir="rtl">ipsum dolor sit</div> ame|t^t');

// rtl forward character
testIt('rtl', 0, 'forward', '^L|orem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 1, 'forward', 'L^o|rem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 2, 'forward', 'Lo^r|em <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 3, 'forward', 'Lor^e|m <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 4, 'forward', 'Lore^m| <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 5, 'forward', 'Lorem ^<div dir="rtl">|ipsum dolor sit</div> amett');
testIt('rtl', 6, 'forward', 'Lorem ^<div dir="rtl">|ipsum dolor sit</div> amett');
testIt('rtl', 7, 'forward', 'Lorem <^d|iv  dir="rtl">ipsum dolor sit amett');
testIt('rtl', 8, 'forward', 'Lorem <d|iv dir="rtl">ipsum dolor sit amett</d|iv>');
testIt('rtl', 9, 'forward', 'Lorem <di|v dir="rtl">ipsum dolor sit amett</di|v>');
testIt('rtl', 10, 'forward', 'Lorem <div| dir="rtl">ipsum dolor sit amett</div|>');
testIt('rtl', 11, 'forward', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 12, 'forward', 'Lorem <div |dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 13, 'forward', 'Lorem <div d|ir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 14, 'forward', 'Lorem <div di|r="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 15, 'forward', 'Lorem <div dir|="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 16, 'forward', 'Lorem <div dir="|&quot;rtl&quot;">ipsum dolor sit</div> amett');
testIt('rtl', 17, 'forward', 'Lorem <div dir="|rtl">ipsum dolor sit</div> amett');
testIt('rtl', 18, 'forward', 'Lorem <div dir="r|tl">ipsum dolor sit</div> amett');
testIt('rtl', 19, 'forward', 'Lorem <div dir="rt|l">ipsum dolor sit</div> amett');
testIt('rtl', 20, 'forward', 'Lorem <div dir="rtl|">ipsum dolor sit</div> amett');
testIt('rtl', 21, 'forward', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 22, 'forward', 'Lorem <div dir="rtl">^i|psum dolor sit</div> amett');
testIt('rtl', 23, 'forward', 'Lorem <div dir="rtl">i^p|sum dolor sit</div> amett');
testIt('rtl', 24, 'forward', 'Lorem <div dir="rtl">ip^s|um dolor sit</div> amett');
testIt('rtl', 25, 'forward', 'Lorem <div dir="rtl">ips^u|m dolor sit</div> amett');
testIt('rtl', 26, 'forward', 'Lorem <div dir="rtl">ipsu^m| dolor sit</div> amett');
testIt('rtl', 27, 'forward', 'Lorem <div dir="rtl">ipsum^ |dolor sit</div> amett');
testIt('rtl', 28, 'forward', 'Lorem <div dir="rtl">ipsum ^d|olor sit</div> amett');
testIt('rtl', 29, 'forward', 'Lorem <div dir="rtl">ipsum d^o|lor sit</div> amett');
testIt('rtl', 30, 'forward', 'Lorem <div dir="rtl">ipsum do^l|or sit</div> amett');
testIt('rtl', 31, 'forward', 'Lorem <div dir="rtl">ipsum dol^o|r sit</div> amett');
testIt('rtl', 32, 'forward', 'Lorem <div dir="rtl">ipsum dolo^r| sit</div> amett');
testIt('rtl', 33, 'forward', 'Lorem <div dir="rtl">ipsum dolor^ |sit</div> amett');
testIt('rtl', 34, 'forward', 'Lorem <div dir="rtl">ipsum dolor ^s|it</div> amett');
testIt('rtl', 35, 'forward', 'Lorem <div dir="rtl">ipsum dolor s^i|t</div> amett');
testIt('rtl', 36, 'forward', 'Lorem <div dir="rtl">ipsum dolor si^t|</div> amett');
testIt('rtl', 37, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit^</div>| amett');
testIt('rtl', 38, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit<^/|div> amett</div>');
testIt('rtl', 39, 'forward', 'Lorem <div dir="rtl">ipsum dolor sitdiv^ |amett</div>');
testIt('rtl', 40, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('rtl', 41, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('rtl', 42, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('rtl', 43, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit</div> ^a|mett');
testIt('rtl', 44, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit</div> ^a|mett');
testIt('rtl', 45, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit</div> a^m|ett');
testIt('rtl', 46, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit</div> am^e|tt');
testIt('rtl', 47, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit</div> ame^t|t');
testIt('rtl', 48, 'forward', 'Lorem <div dir="rtl">ipsum dolor sit</div> amet^t|');

// rtl backward character
testIt('rtl', 0, 'backward', '|Lorem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 1, 'backward', '|L^orem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 2, 'backward', 'L|o^rem <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 3, 'backward', 'Lo|r^em <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 4, 'backward', 'Lor|e^m <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 5, 'backward', 'Lore|m^ <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 6, 'backward', 'Lore|m^ <div dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 7, 'backward', 'Lorem |<^div  dir="rtl">ipsum dolor sit amett');
testIt('rtl', 8, 'backward', 'Lorem <d|iv dir="rtl">ipsum dolor sit amett</d|iv>');
testIt('rtl', 9, 'backward', 'Lorem <di|v dir="rtl">ipsum dolor sit amett</di|v>');
testIt('rtl', 10, 'backward', 'Lorem <div| dir="rtl">ipsum dolor sit amett</div|>');
testIt('rtl', 11, 'backward', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 12, 'backward', 'Lorem <div |dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 13, 'backward', 'Lorem <div d|ir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 14, 'backward', 'Lorem <div di|r="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 15, 'backward', 'Lorem <div dir|="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 16, 'backward', 'Lorem <div dir="|&quot;rtl&quot;">ipsum dolor sit</div> amett');
testIt('rtl', 17, 'backward', 'Lorem <div dir="|rtl">ipsum dolor sit</div> amett');
testIt('rtl', 18, 'backward', 'Lorem <div dir="r|tl">ipsum dolor sit</div> amett');
testIt('rtl', 19, 'backward', 'Lorem <div dir="rt|l">ipsum dolor sit</div> amett');
testIt('rtl', 20, 'backward', 'Lorem <div dir="rtl|">ipsum dolor sit</div> amett');
testIt('rtl', 21, 'backward', 'Lorem <div | dir="rtl">ipsum dolor sit</div> amett');
testIt('rtl', 22, 'backward', 'Lorem |<div dir="rtl">^ipsum dolor sit</div> amett');
testIt('rtl', 23, 'backward', 'Lorem <div dir="rtl">|i^psum dolor sit</div> amett');
testIt('rtl', 24, 'backward', 'Lorem <div dir="rtl">i|p^sum dolor sit</div> amett');
testIt('rtl', 25, 'backward', 'Lorem <div dir="rtl">ip|s^um dolor sit</div> amett');
testIt('rtl', 26, 'backward', 'Lorem <div dir="rtl">ips|u^m dolor sit</div> amett');
testIt('rtl', 27, 'backward', 'Lorem <div dir="rtl">ipsu|m^ dolor sit</div> amett');
testIt('rtl', 28, 'backward', 'Lorem <div dir="rtl">ipsum| ^dolor sit</div> amett');
testIt('rtl', 29, 'backward', 'Lorem <div dir="rtl">ipsum |d^olor sit</div> amett');
testIt('rtl', 30, 'backward', 'Lorem <div dir="rtl">ipsum d|o^lor sit</div> amett');
testIt('rtl', 31, 'backward', 'Lorem <div dir="rtl">ipsum do|l^or sit</div> amett');
testIt('rtl', 32, 'backward', 'Lorem <div dir="rtl">ipsum dol|o^r sit</div> amett');
testIt('rtl', 33, 'backward', 'Lorem <div dir="rtl">ipsum dolo|r^ sit</div> amett');
testIt('rtl', 34, 'backward', 'Lorem <div dir="rtl">ipsum dolor| ^sit</div> amett');
testIt('rtl', 35, 'backward', 'Lorem <div dir="rtl">ipsum dolor |s^it</div> amett');
testIt('rtl', 36, 'backward', 'Lorem <div dir="rtl">ipsum dolor s|i^t</div> amett');
testIt('rtl', 37, 'backward', 'Lorem <div dir="rtl">ipsum dolor si|t^</div> amett');
testIt('rtl', 38, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit|<^/div> amett</div>');
testIt('rtl', 39, 'backward', 'Lorem <div dir="rtl">ipsum dolor si|t^div amett</div>');
testIt('rtl', 40, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('rtl', 41, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('rtl', 42, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit amett</div>');
testIt('rtl', 43, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit|</div>^ amett');
testIt('rtl', 44, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit|</div>^ amett');
testIt('rtl', 45, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit</div> |a^mett');
testIt('rtl', 46, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit</div> a|m^ett');
testIt('rtl', 47, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit</div> am|e^tt');
testIt('rtl', 48, 'backward', 'Lorem <div dir="rtl">ipsum dolor sit</div> ame|t^t');
</script>
